requirements.txt
anthropic>=0.8.0
mcp-python>=0.4.0
python-dotenv>=1.0.0
rich>=13.7.0 